Added *.pyc to gitignore
Changes to script to remove dependence on extracted RPX converted to ELF in order to read command lengths.
If a new patch comes out, the file will be needed to update commands.py with any added commands, but its not necessary for general use of the file.
The only files that are needed now are the fighter/ and param/ folders from the patch. Not having the base game available as 0/ means no animation names though.

Also, fixed a bug that was making dumps not happen for subcharacters like projectiles.